orchestrator/requirements.txt
pyyaml>=6.0
python-telegram-bot==13.15
